Add CreateTransactionTallyPatchService (#254)
When we add a transaction to a bill run we also update 'tally' values in the related bill run, invoice and licence records.

We were already aware of duplication across the `BillRunService`, `InvoiceService` and `LicenceService` for handling this. Now, as part of performance improvements identified in [Spike - Investigate deadlock issues with bill runs](https://github.com/DEFRA/sroc-charging-module-api/pull/247) we need to change the way it's done.

So, we're taking the opportunity to put the functionality into a new service that can then be reused. The new service needs to generate a 'patch' object that can be used with an Objection model's static `query().patch()` method. We have found updating a model in this way, versus using the instance `$patch()` method, gives a slight performance boost.

This change just adds the new service and associated tests. We'll update the services in later PR's as we move them over to the new method of updating the DB.

* Notes 

It's important what we pass through as the patch is right so we check every detail of the patch for all 6 scenarios in our tests. In case we forget why we have so many tests for a relatively small service.

- debit
- debit subject to minimum charge
- credit
- credit subject to minimum charge
- zero value
- zero value subject to minimum charge