Handle call to /generate for generated bill run (#260)
https://trello.com/c/TNehNlxK

This came up as an edge case we had not considered when generating bill runs. What should we do if a request to `/generate` is received for a bill run that has a status of 'generated'?

Whilst we believe running the process for a second time would not change the generated bill run values, we've decided it would be more prudent to just not allow this. We certainly cannot think of a scenario where this would be needed so there is no question regards needing to support it.

So, this change adds a new check to `GenerateBillRunValidationService`. If the `status` is generated we will return a `409` in the same way we do if the status is `generating`.