Re-write unit tests
What the service now returns is so different it required a complete re-write of the tests. The good news is it means we could drop any interaction with the database making them much simpler and cleaner.

We haven't gone into exhaustive detail about what is returned as we feel `create_transaction_tally.service.test.js` has everything covered.