Rename & refactor licence service to use patch
> Based on performance improvements we found in [Spike - Investigate deadlock issues with bill runs](https://github.com/DEFRA/sroc-charging-module-api/pull/247)

In order to provide a performance improvement to the create transaction process we have spotted that using [Objection's](https://vincit.github.io/objection.js/) static `MyModel.query().patch()` performs a little better than the instance `$patch()` version. The key difference is in the static version we set exactly what will be updated. In the instance version, whatever properties are currently populated in the model get updated. If that instance is the result of a previous `SELECT` that will mean everything.

We just created the [CreateTransactionTallyPatchService](https://github.com/DEFRA/sroc-charging-module-api/pull/254) so the refactoring is based on using that to return a 'patch' object, rather than an updated instance of `LicenceModel`.

Finally, `LicenceService` has never been the most expressive of names! So, with this major refactoring, we are also taking the opportunity to rename it.