Re-write unit tests
What the service now returns is so different it required a complete re-write of the tests.

We haven't gone into exhaustive detail about what is returned as we feel `create_transaction_tally.service.test.js` has everything covered.

Worth also noting that the constraint between licence and invoice requires us to create an actual invoice record else none of the tests pass. This is not he case for invoice to bill run. That constraint is fine for us to generate random UUID's as the bill run ID.