Add NextTransactionNumberService and migration
https://trello.com/c/kU8RWnKq

As part of implementing the `PATCH /v2/{regimeId}/bill-runs/{billRunId}/send` endpoint we need to generate transation references. These are generated and assigned to each invoice in a bill run prior to the transaction file for SSCL being generated.

Just like `billRunNumber` each regime and region has their own sequence of numbers. So, we need to record what the last one generated was. It would also be handy if the service returned the fully formatted number (there are other values that make up the reference, not just the sequence number).

So, this change adds the new service and DB migration.