Rename & refactor licence service to use patch (#263)
> Based on performance improvements we found in [Spike - Investigate deadlock issues with bill runs](https://github.com/DEFRA/sroc-charging-module-api/pull/247)

In order to provide a performance improvement to the create transaction process we have spotted that using [Objection's](https://vincit.github.io/objection.js/) static `MyModel.query().patch()` performs a little better than the instance `$patch()` version. The key difference is in the static version we set exactly what will be updated. In the instance version, whatever properties are currently populated in the model get updated. If that instance is the result of a previous `SELECT` that will mean everything.

We just created the [CreateTransactionTallyPatchService](https://github.com/DEFRA/sroc-charging-module-api/pull/254) so the refactoring is based on using that to return a 'patch' object, rather than an updated instance of `LicenceModel`.

Finally, `LicenceService` has never been the most expressive of names! So, with this major refactoring, we are also taking the opportunity to rename it.

* Notes - Re-write unit tests

What the service now returns is so different it required a complete re-write of the tests.

We haven't gone into exhaustive detail about what is returned as we feel `create_transaction_tally.service.test.js` has everything covered.

Worth also noting that the constraint between licence and invoice requires us to create an actual invoice record else none of the tests passes. This is not the case for invoice to bill run. That constraint is fine for us to generate random UUID's as the bill run ID.