https://trello.com/c/kU8RWnKq
As part of implementing the `PATCH /v2/{regimeId}/bill-runs/{billRunId}/send` endpoint we need to generate transaction file references i.e. the file name that will we used when a bill run is exported. This is generated and assigned to the bill run prior to the transaction file for SSCL being generated.

Just like `billRunNumber` each regime and region has its own sequence of numbers. So, we need to record what the last one generated was. It would also be handy if the service returned the fully formatted file reference (there are other values that make up the reference, not just the sequence number).

So, this change adds the new service and DB migration.