Add sequence counters migration
Adds a new `file_number` column to the table. In Version 1 it was originally created with a default of 0. But at some point Tony hard coded adding 50000 to the result. We can avoid this by just defaulting the column to 50000 when first created

```
this.file_reference = 50000 + await this.sequenceCounter.nextFileNumber()
```

https://github.com/DEFRA/charging-module-api/blob/main/app/schema/pre_sroc/wrls/bill_run.js#L156