Make filename prefix configurable
In version 1 the prefix is hard coded to `nal`. This appears to come from the fact WRLS are the only regime currently using the API and their current system is NALD.

It looks like it was hard coded for expediency, but if it's regime dependent it should really be based on that.

Without creating a whole new object, the best place to add the values was the existing `rules_service.config.js`. We may want to review our naming for this to something more generic (`regime.config.js` ??) in the future.