Add SendBillRunPresenter
https://trello.com/c/kU8RWnKq

As part of implementing the PATCH `/v2/{regimeId}/bill-runs/{billRunId}/send` endpoint we need to add a presenter that will handle the response to the user.

This endpoint will result in a file reference being generated. Unlike our other `PATCH` requests we want to return the file reference as a response. So we need a presenter to manage that.