Customer Details controller and endpoint (#266)
https://trello.com/c/xGdyLwoK/1926-m-develop-change-customer-details-v2

To complete Change Customer Details, we create a controller and endpoint and connect the service to these.

Note that a small change was made to CreateCustomerDetailsService to change the findOrInsert method to the insert method. This mirrors the behaviour of the v1 endpoint, which simply adds new customer changes to the table regardless of what entries already exist.