Add NextFileReferenceService and migration (#270)
https://trello.com/c/kU8RWnKq

As part of implementing the `PATCH /v2/{regimeId}/bill-runs/{billRunId}/send` endpoint we need to generate transaction file references i.e. the file name that will we used when a bill run is exported. This is generated and assigned to the bill run prior to the transaction file for SSCL being generated.

Just like `billRunNumber` each regime and region has its own sequence of numbers. So, we need to record what the last one generated was. It would also be handy if the service returned the fully formatted file reference (there are other values that make up the reference, not just the sequence number).

So, this change adds the new service and DB migration.

* Notes

** Default counter to 50000

In Version 1 it was originally created with a default of 0. But at some point, Tony hardcoded adding 50000 to the result. We can avoid this by just defaulting the column to 50000 when first created.

** Make filename prefix configurable

In version 1 the prefix is hardcoded to `nal`. This appears to come from the fact WRLS are the only regime currently using the API and their current system is NALD.

It looks like it was hardcoded for expediency, but if it's regime dependent it should really be based on that.

Without creating a whole new object, the best place to add the values was the existing `rules_service.config.js`. We may want to review our naming for this to something more generic (`regime.config.js` ??) in the future.