Add DB trans. param to NextFileReferenceService
As everything to do with updating the bill run and invoices will be managed in our new `SendBillRunReferenceService` we want to do it in a transaction. This includes updating the sequence counters table to protect against incrementing to a new reference and then not using it because of an error.

We've check with the unit tests and they all still pass when `trx` is null. Also, though we normally avoid short param names `trx` is our convention throughout the code base for an Objection DB transaction so it made sense to stick with it here.