Automatically determine deminimis and zero value
To support testing of the SendBillRunReferenceService we need to be able to identify invoices that are 'billable'.

A billable invoice is one that is not flagged as 'deminimis' or 'zero value'. We intend to add a modifier to the `InvoiceModel` which uses these flags in the `invoices` table to determine if an invoice is billable.

To support this it will be a big help if the test `InvoiceHelper` could set the flags automatically based on the values passed in.

With this change it will. Admittedly it ain't the prettiest of code but it is easy to follow and matches the existing modifiers in the `InvoiceModel`.