Add test for no billable invoices scenario
It is possible that none of the invoices in a bill run are 'billable'. In this case we still want to process the bill run (it will eventually get a status of 'billing_not_required`) but we don't want to generate a transaction reference.

If we did the files we send to SSCL would have gaps in their references and this could lead to confusion and concern. Did they miss a file? Has one got lost in transit?