Hook up /send bill run endpoint to new service
https://trello.com/c/kU8RWnKq

This is the final step in implementing the `PATCH /v2/{regimeId}/bill-runs/{billRunId}/send` endpoint.

We just need to hook up our standin endpoint to the new `SendBillRunReferenceService`.

Note - In a future iteration we will also trigger creating of the export file as a background task. However, there is additional work that needs to be done first before this can happen. The scope of the endpoint right now is to get the bill run ready for being exported.