Add SendBillRunReferenceService (#272)
https://trello.com/c/kU8RWnKq

As part of implementing the PATCH /v2/{regimeId}/bill-runs/{billRunId}/send endpoint we need a service that manages generating transaction and file references and applying them to the invoices and bill run. It also needs to be able to handle when no invoices are to be billed.

This change adds the service.

* Notes

** Add DB trans. param to NextFileReferenceService and NextTransactionReferenceService

As everything to do with updating the bill run and invoices will be managed in our new `SendBillRunReferenceService` we want to do it in a transaction. This includes updating the sequence counters table to protect against incrementing to a new reference and then not using it because of an error.

We've checked with the unit tests and they all still pass when `trx` is null. Also, though we normally avoid short param names `trx` is our convention throughout the code base for an Objection DB transaction so it made sense to stick with it here.

** Add new $approved() method to BillRunModel

We are going to use this in our new service as part of the validation. Only 'approved' bill runs can be sent.

** Add billable() modifier to InvoiceModel

Only invoices which will be exported to SSCL should have transaction references generated. This adds a modifier that allows us to easily query for invoices that fit this requirement. To support this we updated `InvoiceHelper` so that it could set the flags like `deminimisInvoice` automatically based on the values passed in.