Update to use UPSERT()
Postgresql supports the concept of an [UPSERT](https://www.postgresql.org/docs/12/sql-insert.html#SQL-ON-CONFLICT), an 'UPDATE or INSERT' query.

You write the query with the intention it should insert a new record. If that fails due to a constraint, in our case a composite key, it will then do what is defined in `ON CONFLICT`.

> `ON CONFLICT DO UPDATE` guarantees an atomic `INSERT` or `UPDATE` outcome; provided there is no independent error, one of those two outcomes is guaranteed, even under high concurrency.

It is our hope that this resolves the issue we see when running the WRLS teams 'cm-test-script'. When job concurrency is set to 4 we see 3 of the first 4 requests fail because all 4 are attempting to insert the invoice record for the first time. When we run the script with this change _all_ requests succeed.

It should be noted there are issues. There is no `upsert()` method in [Objection](https://github.com/Vincit/objection.js/issues/1406). Knex does support it with its [onConflict() method](http://knexjs.org/#Builder-onConflict). But support is limited. We need to increment existing values in the event of conflict and there doesn't appear to be a way to tell knex to do this.

So, we have had to resort to creating our own query using `Knex.raw()`. Also, the change included here is our first pass 'get it working' version. If this is confirmed as a solution to the `DEADLOCK` issue we'll review and come back with a cleaner solution.