Delete Bill Run service (#268)
https://trello.com/c/IKn83UpO/1927-s-develop-delete-bill-run-v2

The next step of developing Delete Bill Run is to create the DeleteBillRunService, which is done in this change.

Note that a small change is made to the migration db/migrations/20201203144442_create_transactions.js to remove the foreign key constraint on transactions.bill_run_id which we've since decided is unnecessary. We edited this migration rather than create a new one due to the difficulty in knex's inability to amend an existing constraint -- we are happy that we can do this at present since the service has not yet been deployed live.