Add 'pending' as an uneditable bill run state
https://trello.com/c/kU8RWnKq

We have recently added support for 'sending' a bill run, which means telling the service it can be sent to SSCL for billing.

At the moment we update the status of the bill run to `pending` when a `/send` request comes in. The intent is it will then be exported to a transaction file and marked as `billed` when complete.

But from the client system's perspective `pending` is equivalent to `billed` as the export could happen anytime after the the `/send` request has been processed.

This change add's `pending` to the uneditable states to ensure we don't allow any changes to happen to the bill run which could result in a discrepency between the exported file and the data in our DB.