Refactor soltuion to be consistent
This updated solution is provided by @StuAA78. It ensures that what we do for `transactionType` is consistent with what we do for `signedChargeValue`.

He also highlighted that by using `{ ...invoice, }` the object we get back is no longer an instance of `InvoiceModel` whereas using `Object.assign` keeps the type intact.