Add new CreateTransactionUpsertService
This is the new service that will eventually replace

- CreateTransactionBillRunService
- CreateTransactionInvoiceService
- CreateTransactionLicenceService
- CreateTransactionTallyService

For the invoices and licences we need to generate an UPSERT statement. For bill runs we just need the UPDATE statement. The problem for all 3 is which fields get updated are based on the transaction to be created

- is a debit, credit, or zero value?
- is it subject to minimum charge?

These questions need to be asked for all 3 and depending on the answers, different combinations of 'tally' fields need to get updated.

The result also needs to be applied in 3 places

- in an `INSERT` statement in case the record is new
- in an `UPSERT ON CONFLICT` statement in the case of invoices and licences
- in an `UPDATE` statement in the case of bill runs

So, we have created this service to handle doing this for all 3 record types in one place. On first looks it is a 'busy' service. But in essence it's

- making a determination about the transaction, for example, is it a credit?
- generating the relevant `credit` statements for all 3 situations

It's admittedly not the clearest of services. But at this time we are struggling to think of an alternative that wouldn't result in duplication of the questions, statement generation, or both.