Fix deminimis/minimum charge defect (#286)
It was discovered that deminimis was not being correctly applied in some scenarios involving minimum charge, eg:
- Debit subject to minimum charge with charge value of 1 is added to an invoice.
- Credit with charge value of 2500 is added to the same invoice so the total value of the invoice is 2499.
- The bill run is generated.
- The debit value of the invoice is therefore adjusted to 2500 so the total value of the bill run is now 0.

Since this falls into deminimis, the invoice should not be included in the bill run. However we were not including invoices with minimum charge values when determining whether an invoice was subject to deminimis and therefore the invoice would be included. This change resolves this by amending InvoiceModel to no longer exclude minimum charge invoices from the deminimis selection.

We also update breaking unit tests:
- InvoiceModel tests are changed to use realistic values for minimum charge invoices;
- A GenerateBillRunService test is changed to correctly expect no invoice in this scenario.