Add supporting methods to Invoice and LicenceModel
To generating the `UPSERT` statements we need to know specifics about the record we are 'upserting'. What is the name of the fields that will cause the `ON CONFLICT` constraint to fire? What fields need to be populated in the `INSERT` statement?

The `CreateTransactionUpsertService` won't know this. But it will know which record type it is creating it for via its `modelClass` param. Rather than have the service, for example, determine the fields used in the constraint for each class, it just asks the class to tell it.