Move to a "tell don't ask" design
On reviewing what we had done so far we realised we were breaking the object-orientation principle of [Tell don't ask](https://martinfowler.com/bliki/TellDontAsk.html) in a big way.

Our `CreateTransactionUpsertService` was asking the model class a lot of questions in order to carry out its work. So, wouldn't it be better if we just 'told' the model class to upsert?

Yes! By switching to this design we reduce the dependencies between the models and the `CreateTransactionUpsertService`, plus simplify the code elsewhere. for example `CreateTransactionService`.

The  model classes know better what statement needs to be generated to allow the upsert to happen, so moving the logic there is a much better fit.

After making this change we then realised `CreateTransactionUpsertService` was no longer 'upserting'. In fact, it was just a duplicate of our defunt service `CreateTransactionTallyService`, but with knobs on! So, the other change we have made in this commit is to copy what remained of `CreateTransactionUpsertService` into `CreateTransactionTallyService`, and then ditch `CreateTransactionUpsertService`.

We've also made some other tweaks like, for example, just passing the table name through rather than the whole class as this again removes the need for the tally service to ask questions of the model class.