Add validation check to controller
We've broken convention here. Typically we do all work in a single service which we then hook up to the controller. In this case we do the bill run validation check outside the service.

We do this for 2 reasons

- there is already a lot going on in the `CreateTransactionService` so this helps reduce the number of moving parts within it
- we have an expectation that we will need to do other work in the controller, to reset the bill run when new transactions are added to a 'generated' bill run. The validation check will need to take place prior to us doing any work

So, to cover this we also add an additional test for the controller to confirm this check is done.