Add tests for Invoice transactionTallyUpsert()
Add tests to cover our new static `transactionTallyUpsert()` on the `InvoiceModel`.

Something we didn't expect was to get an error when passing `null` as the transaction instance. When we have done this with Objection it just seems to handle it. Knex on the overhand errors. To avoid creating, starting, and then committing Knex transactions in our tests we've had to add a method to the `InvoiceModel` to handle null transactions.