Add tests LicenceModel transactionTallyUpsert()
Same as for `InvoiceModel` we add tests to cover `transactionTallyUpsert()`. And just the same as the `InvoiceModel` we had to add an `_applyUpsert()` method to handle the transaction object.

This is a target for possible refactoring between the 2 model classes; they share very similar properties so is their a way we can reuse/share them rather than duplicate.