Add more data overrides to test LicenceHelper (#291)
We will need the ability to set initial values of 'tally' fields for licences when testing. This is to support testing we'll be doing as we move from a 'find or insert' pattern to one that uses [UPSERT (on conflict)](https://www.postgresql.org/docs/current/sql-insert.html).

This change expands what we can set in the `LicenceHelper`.