Add base model that provides UPSERT support
https://trello.com/c/1cZkVwM1

We know we need to start UPSERTing with the invoice and licence models. The code to do this is nearly identical for both. We could shift it off into a service but then we're back to a thing asking the models lots of questions. So, to avoid that and any duplication we are going with making a new base model class that `InvoiceModel` and `LicenceModel` will inherit from. The new model class will inherit from `BaseModel` so that invoice and licence still have access to everything the other models have.

This change adds the new model and upsert related code but holds off actually calling it. We'll switch over to using the new UPSERT methods in a subsequent change.