Update CreateTransactionTallyService for UPSERT
https://trello.com/c/1cZkVwM1

In order to use a PostgreSQL `UPSERT` command we are going to need to build our own SQL statement. An `UPSERT` statement is

```
  INSERT INTO invoices ([field names]) VALUES ([values to be inserted])
  ON CONFLICT ([constraint field names]) DO UPDATE
  SET [field name] = EXCLUDED.[field name]
```

Basically, we are going to need to create both an 'INSERT' and an 'UPDATE' query. In our case what we are inserting or updating is based on the transaction. Is it a credit or a debit? Is it zero value? Is it subject to minimum charge?

Our current design assumes we'll be patching the invoice and licence records and we use the `CreateTransactionTallyService` to generate the 'patch'. This change builds (explodes!) on the existing functionality in the service to also handle putting together what we need for those insert and update statements. We still just need to 'patch' the bill run when a transaction comes in which means we need to keep what we already have.

So, this change

- expands the functionality of `CreateTransactionTallyService` ready for the switch to `UPSERT`
- ensures thes the existing functionality can still access the 'patch' object the service generates