Add new interim BaseUpsertModel
We'll provide upsert functionality to both the invoice and licence models through a new base class. But they still need the functionality `BaseModel` provides. So, this change adds a new interim `BaseUpsertModel` class and has invoice and licence extend that.

We don't do anything else as we're just confirming this doesn't break anything.