Implement base upsert functionality
Add to the `BaseUpsertModel` our shared UPSERT functionality as well as define the methods we expect the extending class to implement.

We also add some tests for the base model's functions to confirm they will throw errors if not overriden. The rest of the base class we expect to be tested through the classes that extend it, which we are comfortable with.