Refactor update bill run when creating transaction
https://trello.com/c/1cZkVwM1

This is another change coming out of our [DEADLOCK solution](https://github.com/DEFRA/sroc-charging-module-api/pull/283) work.

We have now refactored the updating of invoices and licences when a transaction is added to use a [Postgresql UPSERT](https://www.postgresql.org/docs/current/sql-insert.html) statement. As part of this we moved to a "tell don't ask" design where we tell, for example, the `InvoiceModel` to `updateTally(transaction)` itself.

We also [moved the bill run validation](https://github.com/DEFRA/sroc-charging-module-api/pull/297) out into a new `CreateTransactionBillRunValidationService`. This leaves `CreateTransactionBillRunService` not doing much, and doing it differently to `InvoiceModel` and `LicenceModel`.

So, this change is about refactoring the project to bring `BillRunModel` inline with our "tell don't ask" approach.

> For those unaware the upsert pattern will not work for bill run as the `bill_run` record is created outside of the create transaction process.