Refactor to use new `patchTally()` method
The `patchTally()` method is named to be similar, though slightly different to `InvoiceModel.updateTally()`. This is an attempt to both link the 2 methods but also demonstrate there is a difference.

Within the method though we have followed a similar approach; use the `CreateTransactionTallyService` to determine the patch needed and apply the change there rather than just return it.