Add trans. reference to invoice in VIEW bill run
https://trello.com/c/TvlmPtZ4

We've spoken with the WRLS team. They are still knee deep in re-writing their service to talk to V2 of the API so can't say 100% that having the invoice transaction reference in the `GET /bill-runs/{id}` is needed. But having it there might give them more options when it comes to updating their integration.

It's so simple to add, we're doing it in this change. If it turns out not to be needed, we can just as easily remove it.