Add new ViewBillRunInvoicePresenter
This adds a new presenter and then updates ViewBillRunPresenter to use it to manage formatting the invoices. This means we can avoid messing with the query to try and control how it is formatted.

Because we now have some logic in the `ViewBillRunPresenter` we also add a test to help ensure things are as expected.