Fix send bill run link to generate file issue (#312)
https://trello.com/c/NfozGoQ6

We have spotted an issue in how the `SendBillRunReferenceService` is interacting with the `SendTransactionFileService` service via the `BillRunsController` `send()` action.

The `SendTransactionFileService` because the instance of bill run it is seeing doesn't have a status of `pending`. With this change, we ensure the service will see an updated version of the bill run with the correct status.

* Notes - Fix getting file name to SendFileToS3Service

This also fixes some issues we spotted with getting the generated filename to `SendFileToS3Service`.