Fix Docker local permissions and bump Node version (#311)
We encountered an issue where we were unable to run npm install to refresh dependencies in our local container. We tracked this down to the fact that we were enforcing the node non-root user for all environments. To resolve this, we've changed Dockerfile so that this user is only enforced in production, allowing us to do root user stuff in our local dev environment.

While updating Dockerfile we also noticed that the LTS version of Node is now 14, whereas before we were on Node 12. After checking that our unit tests all pass in the new version, we've updated it accordingly.