Add new docker helpers for local development
Working with the API in Docker has been awesome. But we have found there are few things that could make our lives easier

- being able to properly scratch and rebuild the image and environment
- being able to reset the database

The first one especially highlighted an incorrect assumption we had made. We assumed that deleting the image meant it was rebuilt from scratch the next time we ran `docker-compose up`. In fact, we found it was reusing cached layers because as far as it could see nothing had changed.

This changes gives us a couple more VSCode tasks to allow us to rebuild the environment properly, or just refresh the DB for those times we want to reset the data.