Fix generate file write permissions
https://trello.com/c/NfozGoQ6

Having added the ability to generate transaction files and send them to AWS S3, everything was working fine when tested locally. When deployed to our AWS environment though we were seeing `Access Denied` errors.

We have surmised the issue is the difference in the container when run locally versus run in AWS.

Locally we are targetting `development` which allows files to be written in the running docker container. This was irrespective of our recent [Fix Docker local permissions](https://github.com/DEFRA/sroc-charging-module-api/pull/311) as the folder is mounted.

When run in AWS though

- we are running as the `node` user which does not have write permissions
- trying to write to a folder local to the project would require special permissions to be assigned

Most OS's including those in Docker come with an in-built folder specifically for working with temporary files; `/tmp/`.

This change updates the service to default to using it, and only using something local if overriden in the config.