Add credit and debit line values back to invoices
https://trello.com/c/TvlmPtZ4
https://trello.com/c/qldEb08U

Recently we [Removed tally fields from invoices in view bill run](https://github.com/DEFRA/sroc-charging-module-api/pull/313) and from [invoices in view /invoices](https://github.com/DEFRA/sroc-charging-module-api/pull/306).

We had confirmed this with the WRLS team, but an overlooked piece of functionality has been highlighted in recent testing which relies on them.

For annual and two-part tariff bill runs all the underlying invoices are either all credits, or all debits. So, in their UI they don't bother showing the breakdown. But for supplementary bill runs, invoices (we should probably use a more generic name 😁) can have a mix of credit and debit lines. So, in their UI they _do_ like to show the breakdown.

When we removed the fields, we broke the function which grabbed this information for the UI view. This change adds them back into `GET /bill-runs` and `GET /invoices` giving them the flexibility to decide where to grab the info from, and helping keep the `invoice` schema consistent across the 2 responses.