SendTransactionFileService error handling (#315)
After discovering an issue with uncaught errors during the process of sending a transaction file, we reviewed the error handling and decided to make a change.

Previously, SendTransactionFileService was receiving the notify server method (which we use for Airbrake logging) and passing it to each of the services it called: GenerateTransactionFileService and SendFileToS3Service. We did this because the services were intended to run in the background so we wanted to ensure that any issues during the process of generating and sending files were logged using Airbrake instead of disappearing into the ether.

On reflection, we decided it would be better for these services to throw errors, and for SendTransactionFileService to be responsible for catching these and logging them using notify.

Implementing this means we don't have to keep passing notify around to each service. We also no longer need to return boolean values to signal whether or not we succeeded, as this is implicit in the fact that we didn't throw an error!

A further change has been made to GenerateTransactionFileService to properly wrap createWriteStream in a promise as the previous way of doing this was leading to errors being thrown incorrectly.